Improve support for Debian's default compiler flags
Debian adds in extra arguments to the compiler to enable extra security
hardening and debug features. debhelper exports these, which are typically
picked up automatically by autotools, but since we're using our own Makefile,
we'll need to support them manually[1].

We already have a CFLAGS variable, so set it to inherit the environment's
settings.

Rename the LINK variable to the LDFLAGS name that Debian uses, and also set it
to inherit.

CPPFLAGS can simply be appended to CFLAGS, so do that.

This should fix the remaining buildd log scanner warnings[2].

[1] https://wiki.debian.org/HardeningWalkthrough#Handwritten_Makefiles
[2] https://qa.debian.org/bls/packages/p/poolcounter.html

Thanks to Moritz for explaining how the `dpkg-buildflags` system works to me.

Change-Id: Iae262c514ab2ed54e461b9eb3ceb6184b6df7a9c