Add Blubber/Pipeline configuration
CI has a specific image to test poolcounter which inherits the generic
tox image with the addition of libevent and gcc.  I would like to remove
that extra image to:
* ease changes made by developers
* reduce the number of custom images in CI (T397429)

Add a test image in Blubber which installs gcc and libevent-dev from
apt. The entrypoint is tox, as it was in the CI image.

Add `make test-ci` to build and run the test image locally.

Bug: T399075
Depends-On: I1808415b5f023fe04754c7159a769ef840134b96
Change-Id: I8c8a7770e5a94519317e00b0b6b760c16a5bb3a2