Fix flappy lock/unlock test
When a lock is released and clients are waiting on it, poolcounter will
either:
* wake the client by sending them DONE
* immediately attribute the lock (sending them LOCKED)

When a second client sends a lock request and the first client releases
it immediately after, epoll/libevent might process them out of order
(epoll might just round robin them).  When it happens, the lock is
released first and the second clients receives the lock instead of a
wakeup.

I found three affected tests (but there might be more):
* test_second_lock_first_unlock_any
* test_done_no_unlock
* test_done_no_consume_lock

The workaround I found is to add an extra event before releasing the
lock. It is easy enough and does not touch the daemon code.

I have done the debugging using:

  make DEBUG=1
  .tox/py3/bin/pytest -vv -s tests/test_two_clients.py -k test_done_no_unlock

Bug: T399257
Change-Id: Ife54ef9d4ac91549a34b19daf2d2f4904983ed3e